fix(statistics): remove the logging when there is an error caused by creating an existent index
The statistics jobs run in parallel and they try to create the same
index. One of them runs the request to create the index successfully, but
the other one fails due to the index already existing. This error is caught
and logged.

If the error is not related to the index exists, this is thrown and it
is caught and logged in another method related to the statistics job

More information in:
- https://github.com/wazuh/wazuh-kibana-app/issues/4203#issuecomment-1147273268