fix(modules): fixed an error when changing the selected agent and generates a module report.
This problem was caused because the VisHandlers singleton contains the
visualization handlers that are used to get data of the filters applied
and sent to the plugin endpoint. The visualization handlers were removed
of the VisHandlers singleton when changing the selected agent. Due the
dahasboard is not recreated, the visualizations are not added again,
causing that when generating the report, it can't get the applied
filters of the visualizations handlers because the list was empty.