feat(workflow): enhance some workflows.
Create a reusable workflow that uses the prebuit Docker images of
development mode to mount the plugin source code and run a command.
Create a workflow to build a package on demand
Create a workflow to check the linter
Create a workflow to check the unit test with jest. Enhanced to use the
expected platform.
Create a wildcard workflow to run a command in a development mode