fix(compatibility): fix the plugin can't be registered after setup
Moved the request to get the custom logos when the plugin is mounted on
the UI. It means that the plugin is initially registered using the default image
displayed on the platform menu and after the logo can change if it is
customized. It could result in blink to change the logos.