workaround(search-bar): add a filter to the value suggestions in WQL
When getting the distinct values for some fields, the value could not
match the regular expression that validates them, and this causes that
the search can not be run. So, we filters the distinct values to ensure
or reduce the suggestions can be used to search. This causes some
possible values are not displayed in the suggestions. To undone this,
then the API should allow these values.