fix(tag): change the creation of tag in release tool
Sign and annotate the tag created by the tag release tool was moved
to be optional through parameters. This lets its usage if were
required on the future.

The signed/annotated tag causes a problem in the build of Wazuh
dashboard thorugh GHA workflow that uses a reusable workflow to
build the plugins of this repository. This means, the tags that
could be used to build packages of Wazuh dashboard SHOULD NOT
be signed or annotated.