remove(enrollment): remove enrollment.password setting by securiy reasons
The setting is saved in the advanced settings (tenant configuration) and
is visible to each user with read permissions in the tenant. So we
remove the setting to avoid other users can see the value. The user should
indicate the username and password each time the enroll agent assistant is
used.