remove API reference link from initial buffer message
Simplifies the initial buffer by omitting the API reference link,
resulting in a cleaner console message for users.

remove the link to the documentation assumming the dev tools content is stored in the localstorage, this could cause problem when accessing to multiple environment with different versions or on upgrades persisting the previous value.