refactor dashboard components for IT Hygiene Packages Inventories
This commit refactors the dashboard components for IT Hygiene Packages Inventories by introducing the new classes «ITHygienePackagesInventoriesPackagesDashboardLayoutDefinition» and «ITHygienePackagesInventoriesPackagesDashboardByRendererConfig». The previous implementation has been replaced to streamline the dashboard panel creation process and improve maintainability. The new structure enhances the organization of the dashboard layout and configuration, ensuring better integration with the existing dashboard framework.