update visualization for IT Hygiene Packages Inventories Hotfixes dashboard
This update modifies the visualization settings for the IT Hygiene Packages Inventories Hotfixes dashboard. The changes enhance the display of the most common knowledge base articles (KBs) by adjusting the visualization parameters and ensuring that the dashboard accurately reflects the latest data. This will improve user experience and data interpretation for users accessing the dashboard.